End to end for reading the catalog working.

+++ database/src/main/resources/db/migration/0.0.1/database/V0.0.1.16.12.20__ddl_setup.sql
@@ -10,8 +10,8 @@ CREATE TABLE IF NOT EXISTS catalog (
     id INT NOT NULL AUTO_INCREMENT,
     category VARCHAR(50) NOT NULL,
     name VARCHAR(50) NOT NULL,
-    retail_price FLOAT(3,2) DEFAULT 0.00 NOT NULL,
-    wholesale_price FLOAT(3,2) DEFAULT 0.00 NOT NULL,
+    retail_price FLOAT(6,2) DEFAULT 0.00 NOT NULL,
+    wholesale_price FLOAT(6,2) DEFAULT 0.00 NOT NULL,
     PRIMARY KEY (id)
 );
